test: split css minimizer tests into a separate file excluded by jest config
`it.skip`-ing the css-minimizer rows still left orphan snapshots in the
shared snapshot file, which made some CI rows fail when those snapshots
weren't reconciled. Move all of them into a dedicated
`test/css-minify-option.test.js` (with its own snapshot file) and add
a `testPathIgnorePatterns` entry in `jest.config.js` that drops that
file entirely on rows where the CSS deps don't run — Node <18 or
Windows. Removes the now-unneeded continue-on-error / fail-fast
overrides from the workflow.
