+++ test/minify-option.test.js
@@ -1,4 +1,3 @@
-/* eslint-disable jest/no-standalone-expect */
 import path from "path";
 
 import TerserPlugin from "../src";
@@ -11,16 +10,6 @@ import {
   getWarnings,
   readsAssets,
 } from "./helpers";
-
-// Some bundled minimizers require a newer Node than the test matrix
-// covers (`@swc/css` >=14, `cssnano@7` >=18, `esbuild@0.27` >=18, etc.).
-// On top of that, `@swc/css`, `lightningcss`, `esbuild`, and the
-// cssnano native-binary deps don't reliably install on Windows agents.
-// Skip the rows that exercise them when the environment isn't a fit
-// so snapshot-based assertions don't drift.
-const NODE_MAJOR = Number(process.versions.node.split(".")[0]);
-const IS_WINDOWS = process.platform === "win32";
-const RUN_CSS_TESTS = NODE_MAJOR >= 18 && !IS_WINDOWS;
 
 describe("minify option", () => {
   it("should work", async () => {
@@ -1312,390 +1301,4 @@ describe("minify option", () => {
     expect(getErrors(stats)).toMatchSnapshot("errors");
     expect(getWarnings(stats)).toMatchSnapshot("warnings");
   });
-
-  // cssnano@7 requires Node >=18 — older Node rows install older cssnano via CI.
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work using when the `minify` option is `cssnanoMinify`",
-    async () => {
-      const compiler = getCompiler({
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: TerserPlugin.cssnanoMinify,
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work using when the `minify` option is `cssnanoMinify` and allows to set `cssnano` options",
-    async () => {
-      const compiler = getCompiler({
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: TerserPlugin.cssnanoMinify,
-        minimizerOptions: { preset: ["default", { discardComments: false }] },
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work using when the `minify` option is `cssoMinify`",
-    async () => {
-      const compiler = getCompiler({
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: TerserPlugin.cssoMinify,
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work using when the `minify` option is `cssoMinify` and allows to set `csso` options",
-    async () => {
-      const compiler = getCompiler({
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: TerserPlugin.cssoMinify,
-        minimizerOptions: { comments: false },
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work using when the `minify` option is `cleanCssMinify`",
-    async () => {
-      const compiler = getCompiler({
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: TerserPlugin.cleanCssMinify,
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work using when the `minify` option is `cleanCssMinify` and allows to set `clean-css` options",
-    async () => {
-      const compiler = getCompiler({
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: TerserPlugin.cleanCssMinify,
-        minimizerOptions: { format: "beautify" },
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  // esbuild@0.27 requires Node >=18.
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work using when the `minify` option is `esbuildMinifyCss`",
-    async () => {
-      const compiler = getCompiler({
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: TerserPlugin.esbuildMinifyCss,
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  // `lightningcss` requires Node >=12.
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work using when the `minify` option is `lightningCssMinify`",
-    async () => {
-      const compiler = getCompiler({
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: TerserPlugin.lightningCssMinify,
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  // `@swc/css` requires Node >=14.
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work using when the `minify` option is `swcMinifyCss`",
-    async () => {
-      const compiler = getCompiler({
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: TerserPlugin.swcMinifyCss,
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work when `minify` is an array of functions using `cssnanoMinify`",
-    async () => {
-      const compiler = getCompiler({
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: [
-          TerserPlugin.cssnanoMinify,
-          // Second pass: pass-through that asserts the previous minimizer
-          // produced a string we can keep working with.
-          (data) => {
-            const [[, code]] = Object.entries(data);
-
-            return { code };
-          },
-        ],
-        minimizerOptions: [{ preset: "default" }, {}],
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work and merge source maps when `minify` is an array of `terserMinify` minimizers",
-    async () => {
-      const compiler = getCompiler({
-        devtool: "source-map",
-        entry: path.resolve(__dirname, "./fixtures/minify/es6.js"),
-        output: {
-          path: path.resolve(__dirname, "./dist-terser"),
-          filename: "[name].js",
-          chunkFilename: "[id].[name].js",
-        },
-      });
-
-      new TerserPlugin({
-        minify: [TerserPlugin.terserMinify, TerserPlugin.terserMinify],
-        minimizerOptions: [{ mangle: false }, { mangle: true }],
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work and merge source maps when `minify` mixes `terserMinify` with `uglifyJsMinify`",
-    async () => {
-      const compiler = getCompiler({
-        devtool: "source-map",
-        entry: path.resolve(__dirname, "./fixtures/minify/es6.js"),
-        output: {
-          path: path.resolve(__dirname, "./dist-terser"),
-          filename: "[name].js",
-          chunkFilename: "[id].[name].js",
-        },
-      });
-
-      new TerserPlugin({
-        minify: [TerserPlugin.terserMinify, TerserPlugin.uglifyJsMinify],
-        minimizerOptions: [{ mangle: false }, { mangle: true }],
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work and merge source maps when `minify` is an array of CSS minimizers",
-    async () => {
-      const compiler = getCompiler({
-        devtool: "source-map",
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: [TerserPlugin.cssnanoMinify, TerserPlugin.cssnanoMinify],
-        minimizerOptions: [{ preset: "default" }, { preset: "default" }],
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work and merge source maps when `minify` mixes `terserMinify` with `swcMinify`",
-    async () => {
-      const compiler = getCompiler({
-        devtool: "source-map",
-        entry: path.resolve(__dirname, "./fixtures/minify/es6.js"),
-        output: {
-          path: path.resolve(__dirname, "./dist-terser"),
-          filename: "[name].js",
-          chunkFilename: "[id].[name].js",
-        },
-      });
-
-      new TerserPlugin({
-        minify: [TerserPlugin.terserMinify, TerserPlugin.swcMinify],
-        minimizerOptions: [{ mangle: false }, {}],
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work and merge source maps when `minify` mixes `terserMinify` with `esbuildMinify`",
-    async () => {
-      const compiler = getCompiler({
-        devtool: "source-map",
-        entry: path.resolve(__dirname, "./fixtures/minify/es6.js"),
-        output: {
-          path: path.resolve(__dirname, "./dist-terser"),
-          filename: "[name].js",
-          chunkFilename: "[id].[name].js",
-        },
-      });
-
-      new TerserPlugin({
-        minify: [TerserPlugin.terserMinify, TerserPlugin.esbuildMinify],
-        minimizerOptions: [{ mangle: false }, {}],
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
-
-  // The chain runs through every CSS minimizer; `esbuild` is the
-  // tightest constraint at Node >=18.
-  (RUN_CSS_TESTS ? it : it.skip)(
-    "should work and merge source maps when `minify` mixes CSS minimizers using `cssnano`, `csso`, `cleanCss`, `lightningCss`, `swcCss`, and `esbuild`",
-    async () => {
-      const compiler = getCompiler({
-        devtool: "source-map",
-        entry: path.resolve(__dirname, "./fixtures/css.js"),
-      });
-
-      new TerserPlugin().apply(compiler);
-      new TerserPlugin({
-        test: /\.css(\?.*)?$/i,
-        minify: [
-          TerserPlugin.cssnanoMinify,
-          TerserPlugin.cssoMinify,
-          TerserPlugin.cleanCssMinify,
-          TerserPlugin.lightningCssMinify,
-          TerserPlugin.swcMinifyCss,
-          TerserPlugin.esbuildMinifyCss,
-        ],
-        minimizerOptions: [{ preset: "default" }, {}, {}, {}, {}, {}],
-      }).apply(compiler);
-
-      const stats = await compile(compiler);
-
-      expect(readsAssets(compiler, stats)).toMatchSnapshot("assets");
-      expect(getErrors(stats)).toMatchSnapshot("errors");
-      expect(getWarnings(stats)).toMatchSnapshot("warnings");
-    },
-  );
 });
